# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Petstore API",
        "version": "0.0"
    },
    "paths": {
...
        "/api/v1/animals": {
            "get": {
                "tags": [
                    "Animal"
                ],
                "summary": "Get all animals",
                "operationId": "0b709deff337cdadafc2fa4db0a789e8",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AnimalSummaryCollection"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Animal"
                ],
                "summary": "Add a new animal",
                "operationId": "e1cfcfb74160cc91b3fb818a01bc4064",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreAnimalRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/animals/{animal}": {
            "get": {
                "tags": [
                    "Animal"
                ],
                "summary": "Get the given animal",
                "operationId": "e19e43645410a73ed5ba10a9e3ac6d26",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Animal"
                ],
                "summary": "Update the given animal",
                "operationId": "3ccd62da32c62815a141deb15bd4c96d",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Animal"
                ],
                "summary": "Delete the given animal",
                "operationId": "5a8691ecf90cb852d51d904ffe7c7a0f",
                "parameters": [
                    {
                        "name": "animal",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "409": {
                        "description": "Conflict - A pet exists for the animal type"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Log in",
                "description": "Handle an incoming authentication request.",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "No content",
                        "headers": {
                            "Set-Cookie": {
                                "schema": {
                                    "type": "string",
                                    "example": "laravel_session=0%3D; expires=Wed, 12 Feb 2025 21:56:44 GMT; Max-Age=7200; httponly; samesite=lax"
                                }
                            }
                        },
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Log out",
                "description": "Destroy an authenticated session.",
                "operationId": "logout",
                "responses": {
                    "204": {
                        "description": "Session destroyed",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "scalar": "laravel_session"
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "LoginRequest": {
                "title": "LoginRequest",
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "admin@laravelpetshelter.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "password"
                    }
                },
                "type": "object"
            },
            "StoreAnimalRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Dog",
                        "nullable": false
                    }
                },
                "type": "object"
            },
...
            "AnimalSummaryCollection": {
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AnimalSummary"
                        }
                    }
                },
                "type": "object"
            },
            "AnimalSummary": {
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Cat"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "cookie": {
                "type": "apiKey",
                "name": "laravel_session",
                "in": "cookie"
            }
        }
    },
...
}